+++ CHANGELOG.md
@@ -1,11 +1,24 @@
 # Change Log
 
+## [1.0.2]
+
+-   minor fixes
+-   console methods added
+    -   `console.info`
+    -   `console.error`
+    -   `console.count`
+    -   `console.countReset`
+    -   `console.time`
+    -   `console.timeEnd`
+
+---
+
 ## [1.0.0]
 
 -   basic javascript console features
--   console methods
+-   console methods added
     -   `console.log`
     -   `console.warn`
     -   `console.clear`
--   experimental method implementation
+-   experimental methods implemented
     -   `window.alert`